replaced TarantoolDB with TarantoolDBConnection
