+++ planetmint/backend/__init__.py
@@ -13,4 +13,4 @@ configuration or the ``PLANETMINT_DATABASE_BACKEND`` environment variable.
 
 # Include the backend interfaces
 from planetmint.backend import schema, query  # noqa
-from planetmint.backend.connection import Connection
+from planetmint.backend.connection import connect, Connection
