+++ planetmint/backend/schema.py
@@ -9,7 +9,7 @@ from functools import singledispatch
 import logging
 
 from planetmint.config import Config
-from planetmint.backend.connection import Connection
+from planetmint.backend.connection import connect
 from planetmint.transactions.common.exceptions import ValidationError
 from planetmint.transactions.common.utils import (
     validate_all_values_for_key_in_obj, validate_all_values_for_key_in_list)
@@ -79,7 +79,7 @@ def init_database(connection=None, dbname=None):
             configuration.
     """
 
-    connection = connection or Connection()
+    connection = connection or connect()
     dbname = dbname or Config().get()['database']['name']
 
     create_database(connection, dbname)
