+++ planetmint/backend/tarantool/query.py
@@ -10,14 +10,14 @@ from operator import itemgetter
 
 from planetmint.backend import query
 from planetmint.backend.utils import module_dispatch_registrar
-from planetmint.backend.tarantool.connection import TarantoolDB
+from planetmint.backend.tarantool.connection import TarantoolDBConnection
 from planetmint.backend.tarantool.transaction.tools import TransactionCompose, TransactionDecompose
 from json import dumps, loads
 
 register_query = module_dispatch_registrar(query)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def _group_transaction_by_ids(connection, txids: list):
     txspace = connection.space("transactions")
     inxspace = connection.space("inputs")
@@ -53,7 +53,7 @@ def _group_transaction_by_ids(connection, txids: list):
     return _transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_transactions(connection, signed_transactions: list):
     txspace = connection.space("transactions")
     inxspace = connection.space("inputs")
@@ -86,26 +86,26 @@ def store_transactions(connection, signed_transactions: list):
             assetsxspace.insert(txtuples["asset"])
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_transaction(connection, transaction_id: str):
     _transactions = _group_transaction_by_ids(txids=[transaction_id], connection=connection)
     return next(iter(_transactions), None)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_transactions(connection, transactions_ids: list):
     _transactions = _group_transaction_by_ids(txids=transactions_ids, connection=connection)
     return _transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_metadatas(connection, metadata: list):
     space = connection.space("meta_data")
     for meta in metadata:
         space.insert((meta["id"], meta["data"] if not "metadata" in meta else meta["metadata"]))
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_metadata(connection, transaction_ids: list):
     _returned_data = []
     space = connection.space("meta_data")
@@ -116,7 +116,7 @@ def get_metadata(connection, transaction_ids: list):
     return _returned_data if len(_returned_data) > 0 else None
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_asset(connection, asset):
     space = connection.space("assets")
     convert = lambda obj: obj if isinstance(obj, tuple) else (obj, obj["id"], obj["id"])
@@ -126,7 +126,7 @@ def store_asset(connection, asset):
         print("DUPLICATE ERROR")
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_assets(connection, assets: list):
     space = connection.space("assets")
     convert = lambda obj: obj if isinstance(obj, tuple) else (obj, obj["id"], obj["id"])
@@ -137,7 +137,7 @@ def store_assets(connection, assets: list):
             print(f"EXCEPTION : {ex} ")
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_asset(connection, asset_id: str):
     space = connection.space("assets")
     _data = space.select(asset_id, index="txid_search")
@@ -145,7 +145,7 @@ def get_asset(connection, asset_id: str):
     return _data[0][0] if len(_data) > 0 else []
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_assets(connection, assets_ids: list) -> list:
     _returned_data = []
     for _id in list(set(assets_ids)):
@@ -154,7 +154,7 @@ def get_assets(connection, assets_ids: list) -> list:
     return sorted(_returned_data, key=lambda k: k["id"], reverse=False)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_spent(connection, fullfil_transaction_id: str, fullfil_output_index: str):
     space = connection.space("inputs")
     _inputs = space.select([fullfil_transaction_id, str(fullfil_output_index)], index="spent_search")
@@ -163,7 +163,7 @@ def get_spent(connection, fullfil_transaction_id: str, fullfil_output_index: str
     return _transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_latest_block(connection):  # TODO Here is used DESCENDING OPERATOR
     space = connection.space("blocks")
     _all_blocks = space.select()
@@ -183,7 +183,7 @@ def get_latest_block(connection):  # TODO Here is used DESCENDING OPERATOR
     return block
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_block(connection, block: dict):
     space = connection.space("blocks")
     block_unique_id = token_hex(8)
@@ -195,7 +195,7 @@ def store_block(connection, block: dict):
         space.insert((txid, block_unique_id))
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_txids_filtered(connection, asset_id: str, operation: str = None,
                        last_tx: any = None):  # TODO here is used 'OR' operator
     actions = {
@@ -256,7 +256,7 @@ def _remove_text_score(asset):
     return asset
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_owned_ids(connection, owner: str):
     space = connection.space("keys")
     _keys = space.select(owner, index="keys_search")
@@ -267,7 +267,7 @@ def get_owned_ids(connection, owner: str):
     return _transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_spending_transactions(connection, inputs):
     _transactions = []
 
@@ -280,7 +280,7 @@ def get_spending_transactions(connection, inputs):
     return _transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_block(connection, block_id=[]):
     space = connection.space("blocks")
     _block = space.select(block_id, index="block_search", limit=1)
@@ -295,7 +295,7 @@ def get_block(connection, block_id=[]):
     return {"app_hash": _block[0], "height": _block[1], "transactions": [_tx[0] for _tx in _txblock]}
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_block_with_transaction(connection, txid: str):
     space = connection.space("blocks_tx")
     _all_blocks_tx = space.select(txid, index="id_search")
@@ -307,7 +307,7 @@ def get_block_with_transaction(connection, txid: str):
     return [{"height": _height[1]} for _height in _block.data]
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def delete_transactions(connection, txn_ids: list):
     tx_space = connection.space("transactions")
     for _id in txn_ids:
@@ -335,7 +335,7 @@ def delete_transactions(connection, txn_ids: list):
         assets_space.delete(_id, index="txid_search")
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_unspent_outputs(connection, *unspent_outputs: list):
     space = connection.space('utxos')
     result = []
@@ -345,7 +345,7 @@ def store_unspent_outputs(connection, *unspent_outputs: list):
             result.append(output.data)
     return result
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def delete_unspent_outputs(connection, *unspent_outputs: list):
     space = connection.space('utxos')
     result = []
@@ -356,14 +356,14 @@ def delete_unspent_outputs(connection, *unspent_outputs: list):
     return result
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_unspent_outputs(connection, query=None):  # for now we don't have implementation for 'query'.
     space = connection.space('utxos')
     _utxos = space.select([]).data
     return [loads(utx[2]) for utx in _utxos]
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_pre_commit_state(connection, state: dict):
     space = connection.space("pre_commits")
     _precommit = space.select(state["height"], index="height_search", limit=1)
@@ -375,7 +375,7 @@ def store_pre_commit_state(connection, state: dict):
                  limit=1)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_pre_commit_state(connection):
     space = connection.space("pre_commits")
     _commit = space.select([], index="id_search").data
@@ -385,7 +385,7 @@ def get_pre_commit_state(connection):
     return {"height": _commit[1], "transactions": _commit[2]}
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_validator_set(conn, validators_update: dict):
     space = conn.space("validators")
     _validator = space.select(validators_update["height"], index="height_search", limit=1)
@@ -397,7 +397,7 @@ def store_validator_set(conn, validators_update: dict):
                  limit=1)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def delete_validator_set(connection, height: int):
     space = connection.space("validators")
     _validators = space.select(height, index="height_search")
@@ -405,7 +405,7 @@ def delete_validator_set(connection, height: int):
         space.delete(_valid[0])
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_election(connection, election_id: str, height: int, is_concluded: bool):
     space = connection.space("elections")
     space.upsert((election_id, height, is_concluded),
@@ -415,7 +415,7 @@ def store_election(connection, election_id: str, height: int, is_concluded: bool
                  limit=1)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_elections(connection, elections: list):
     space = connection.space("elections")
     for election in elections:
@@ -424,7 +424,7 @@ def store_elections(connection, elections: list):
                                   election["is_concluded"]))
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def delete_elections(connection, height: int):
     space = connection.space("elections")
     _elections = space.select(height, index="height_search")
@@ -432,7 +432,7 @@ def delete_elections(connection, height: int):
         space.delete(_elec[0])
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_validator_set(connection, height: int = None):
     space = connection.space("validators")
     _validators = space.select()
@@ -446,7 +446,7 @@ def get_validator_set(connection, height: int = None):
         return next(iter(sorted(_validators, key=lambda k: k["height"], reverse=True)), None)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_election(connection, election_id: str):
     space = connection.space("elections")
     _elections = space.select(election_id, index="id_search")
@@ -457,7 +457,7 @@ def get_election(connection, election_id: str):
     return {"election_id": _election[0], "height": _election[1], "is_concluded": _election[2]}
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_asset_tokens_for_public_key(connection, asset_id: str, public_key: str):
     space = connection.space("keys")
     # _keys = space.select([public_key], index="keys_search")
@@ -469,7 +469,7 @@ def get_asset_tokens_for_public_key(connection, asset_id: str, public_key: str):
     return _grouped_transactions
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def store_abci_chain(connection, height: int, chain_id: str, is_synced: bool = True):
     space = connection.space("abci_chains")
     space.upsert((height, is_synced, chain_id),
@@ -479,7 +479,7 @@ def store_abci_chain(connection, height: int, chain_id: str, is_synced: bool = T
                  limit=1)
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def delete_abci_chain(connection, height: int):
     space = connection.space("abci_chains")
     _chains = space.select(height, index="height_search")
@@ -487,7 +487,7 @@ def delete_abci_chain(connection, height: int):
         space.delete(_chain[2])
 
 
-@register_query(TarantoolDB)
+@register_query(TarantoolDBConnection)
 def get_latest_abci_chain(connection):
     space = connection.space("abci_chains")
     _all_chains = space.select().data
